refactor: Update +page.server.js to load review data
This commit updates the `load` function in `+page.server.js` to load review data from the database using the `readReview` function. The retrieved data is then transformed into a serializable format before being returned as props. This change improves the functionality of the page by providing the necessary data for rendering reviews.