chore: Refactor admin page server code and update admin page layout
This commit refactors the server code for the admin page and updates the layout of the admin page. The server code has been modified to include a new data source and serialize the data. The layout of the admin page has been improved by adding styling and organizing the data in separate sections.